Conflicts:
	amp_conf/htdocs/admin/libraries/Composer/composer.json
	amp_conf/htdocs/admin/libraries/Composer/composer.lock
	amp_conf/htdocs/admin/libraries/Composer/vendor/composer/autoload_files.php
	amp_conf/htdocs/admin/libraries/Composer/vendor/composer/installed.json

+++ amp_conf/htdocs/admin/libraries/Console/Stop.class.php
@@ -15,9 +15,12 @@ class Stop extends Command {
 	protected function configure(){
 		$this->setName('stop')
 			->setDescription(_('Stop Asterisk and run other needed FreePBX commands'))
-			->setDefinition(array(
-				new InputOption('immediate', 'i', InputOption::VALUE_NONE, _('Shutdown NOW rather than convieniently')),
-				new InputArgument('args', InputArgument::IS_ARRAY, null, null),));
+			->addOption('pre', null, InputOption::VALUE_NONE, _('Only run pre-stop hooks'))
+			->addOption('post', null, InputOption::VALUE_NONE, _('Only run post-stop hooks'))
+			->addOption('immediate', 'i', InputOption::VALUE_NONE, _('Shutdown NOW rather than convieniently'))
+			->addOption('maxwait', 'm', InputOption::VALUE_OPTIONAL, _('Maximum time (in seconds) to wait for asterisk to stop gracefully. Default 30 seconds'))
+			->addArgument('args', InputArgument::IS_ARRAY, _('Module names'))
+			->setHelp($this->showHelp());
 	}
 
 	protected function execute(InputInterface $input, OutputInterface $output){
@@ -25,138 +28,170 @@ class Stop extends Command {
 			$output->writeln("<error>"._("You need to be root to run this command")."</error>");
 			exit(1);
 		}
+		$options = $input->getOptions();
 		$args = $input->getArgument('args');
-		if(!empty($args)) {
-			$pre = $this->preAsteriskHooks($output,false);
-			$post = $this->postAsteriskHooks($output,false);
-			$aststat = $this->asteriskProcess();
-			$asteriskrunning = ($aststat[0]);
-			$bmo = \FreePBX::create();
-			$found = false;
-			foreach($pre as $pri => $data) {
-				if(strtolower($data['module']) == $args[0]) {
-					$found = true;
-					if(!$asteriskrunning) {
-						$output->writeln("<error>"._('This service must be stopped before Asterisk has stopped')."</error>");
-						break;
-					}
-					$bmo->$data['module']->$data['method']($output);
-					break;
-				}
-			}
-			foreach($post as $pri => $data) {
-				if(strtolower($data['module']) == $args[0]) {
-					$found = true;
-					if($asteriskrunning) {
-						$output->writeln("<error>"._('This service must be stopped after Asterisk has stopped')."</error>");
-						break;
-					}
-					$bmo->$data['module']->$data['method']($output);
-					break;
-				}
-			}
-			if(!$found) {
-				$output->writeln("<error>"._('Unable to find service to start')."</error>");
-			}
-		} else {
-			$brand = \FreePBX::Config()->get("DASHBOARD_FREEPBX_BRAND");
-			$output->writeln(sprintf(_('Running %s shutdown...'),$brand));
-			$output->writeln('');
-			$output->writeln(_('Checking Asterisk Status...'));
-			$aststat = $this->asteriskProcess();
-			if(!$aststat[0]){
-				$output->writeln(_('Asterisk seems NOT to be running'));
-				$output->writeln('<info>'._('Not running Pre-Asterisk Shutdown Hooks.').'</info>');
-			}else{
-				$output->writeln(_('Run Pre-Asterisk Shutdown Hooks'));
-				$this->preAsteriskHooks($output);
-				if($input->getOption('immediate')){
-					$output->writeln(_('Shutting down NOW...'));
-					$this->stopAsterisk($output, now);
-					$aststat = $this->asteriskProcess();
-					sleep(2);
-					while($i++ < 3){
-						if($aststat[0]){
-							$output->writeln(_('The immediate shutdown did not go as planned... Trying again...'));
-							$this->stopAsterisk($output, now);
-							sleep(2);
-						}
-					}
-				}else{
-					$output->writeln('');
-					$output->writeln(_('Shutting down Asterisk Gracefully...'));
-					if (!$output->isQuiet()) {
-						$output->writeln(sprintf(_('Press %s to Cancel'),'C'));
-						$output->writeln(sprintf(_('Press %s to shut down NOW'),'N'));
-					}
-					$progress = new ProgressBar($output, 120);
-					$this->stopAsterisk($output, 'gracefully');
-					$progress->start();
-					$i = 0;
-					if (!$output->isQuiet()) {
-						$stdin = fopen('php://stdin', 'r');
-						stream_set_blocking($stdin,0);
-						exec('stty -g', $term);
-						system("stty -icanon");
-					}
-					while ($stdin) {
-						if (!$output->isQuiet()) {
-							$res = fgetc ($stdin);
-							echo $res;
-						}
-						$aststat = $this->asteriskProcess();
-						if(!$aststat[0]){
-							$progress->finish();
-							$output->writeln('');
-							$output->writeln(_('Asterisk Stopped Successfuly'));
-							$userov = True;
-							break;
-						}
-						if (!$output->isQuiet()) {
-							if (strtolower($res) == 'c'){
-								$progress->finish();
-								$output->writeln('');
-								$this->abortShutdown($output);
-								$userov = True;
-								break;
-							}
-							if (strtolower($res) == 'n'){
-								$progress->finish();
-								$output->writeln('');
-								$this->stopAsterisk($output, 'now');
-								$userov = True;
-								break;
-							}
-						}
-						sleep(1);
-						$progress->advance(1);
-						$i++;
-						if($i == 120){
-							break;
-						}
-					}
-					if (!$output->isQuiet()) {
-						//re-block the stream
-						stream_set_blocking($stdin,1);
-						fclose($stdin);
+		$pre = $this->preAsteriskHooks($output,false);
+		$post = $this->postAsteriskHooks($output,false);
+		$aststat = $this->asteriskProcess();
+		$asteriskrunning = ($aststat[0]);
+		$bmo = \FreePBX::create();
+		$maxwait = (int) $options['maxwait'];
+		if ($maxwait < 5) {
+			$maxwait = 30;
+		}
+		if ($maxwait > 3600) { // 1 hour
+			$maxwait = 3600;
+		}
 
-						system("stty '" . $term[0] . "'");
-					}
-					if(!$userov){
-						$output->writeln(_('Grace Period timed out'));
-						$this->stopAsterisk($output, 'now');
+		// We were asked to only run the pre-stop hooks?
+		if ($options['pre']) {
+			// Note: Do not i18n.
+			$output->writeln("Only running pre-hooks");
+			$runpre = true;
+			$stopasterisk = false;
+			$runpost = false;
+		} elseif ($options['post']) {
+			// Note: Do not i18n.
+			$output->writeln("Only running post-hooks");
+			$runpre = false;
+			$stopasterisk = false;
+			$runpost = true;
+		} else {
+			// Run both
+			$runpre = true;
+			$stopasterisk = true;
+			$runpost = true;
+		}
+
+		// Do we have any params?
+		if ($args) {
+			// We do. Create a temporary array with our hooks, using the ones
+			// we've been asked to do.
+			$stopasterisk = false;
+			$newpre = array();
+			$newpost = array();
+			foreach ($args as $v) {
+				if ($runpre) {
+					foreach ($pre as $pri => $data) {
+						if(strtolower($data['module']) == strtolower($v)) {
+							$newpre[$pri] = $data;
+						}
 					}
 				}
-				$aststat = $this->asteriskProcess();
-				sleep(1);
-				if(!$aststat[0]){
-					$output->writeln('');
-					$output->writeln(_('Running Post-Asterisk Stop Scripts'));
-					$this->postAsteriskHooks($output);
+				if ($runpost) {
+					foreach ($post as $pri => $data) {
+						if(strtolower($data['module']) == strtolower($v)) {
+							$newpost[$pri] = $data;
+						}
+					}
+				}
+			}
+
+			// And overwrite our hooks to run later
+			$pre = $newpre;
+			$post = $newpost;
+		}
+
+		if ($stopasterisk && !$asteriskrunning) {
+			$output->writeln("<error>Asterisk not currently running</error>");
+			$stopasterisk = false;
+		}
+
+		// Now we're ready to go.  
+		$brand = \FreePBX::Config()->get("DASHBOARD_FREEPBX_BRAND");
+
+		$output->writeln(sprintf(_('Running %s shutdown...'),$brand));
+		$output->writeln('');
+
+		if ($runpre) {
+			foreach($pre as $pri => $data) {
+				$bmo->$data['module']->$data['method']($output);
+			}
+		}
+
+		if ($stopasterisk) {
+			if ($options['immediate']) {
+				$output->writeln(_('Shutting down Asterisk Immediately...'));
+				$this->stopAsterisk($output, 'now');
+			} else {
+				$output->writeln(sprintf(_('Shutting down Asterisk Gracefully. Will forcefully kill after %s seconds.'), $maxwait));
+				// Let people force the shutdown if they want
+				$output->writeln(sprintf(_('Press %s to Cancel'),'C'));
+				$output->writeln(sprintf(_('Press %s to shut down NOW'),'N'));
+				// Wait for up to $maxwait before killing it hard
+				$killafter = time() + $maxwait;
+				$starttime = time();
+
+				// Seconds may have ticked over between the two time() calls, which is why
+				// we recalculate.
+				$pct = 100/($killafter - $starttime);
+
+				if (!$output->isQuiet()) {
+					stream_set_blocking(STDIN,0);
+					$term = `stty -g`;
+					system("stty -icanon -echo");
+				}
+
+				$progress = new ProgressBar($output, 100);
+				$this->stopAsterisk($output, 'gracefully');
+				$progress->start();
+				$isrunning = true;
+
+				while ( time() < $killafter ) {
+					if (!$output->isQuiet()) {
+						$res = fread(STDIN,1);
+						if ($res) {
+							if (strtolower($res) === "c") {
+								$progress->setProgress(100);
+								$progress->finish();
+								print "\n";
+								$output->writeln(_('Aborting Shutdown. Asterisk is still running'));
+								$this->abortShutdown($output);
+								system("stty $term");
+								exit(1);
+							} elseif (strtolower($res) === "n") {
+								print "\n";
+								$output->writeln(_('Killing asterisk forcefully.'));
+								$this->stopAsterisk($output, 'now');
+							}
+						}
+					}
+					$current =  (int) (time() - $starttime) * $pct;
+					$progress->setProgress($current);
+					$aststat = $this->asteriskProcess();
+					$asteriskrunning = ($aststat[0]);
+					if (!$asteriskrunning) {
+						$progress->setProgress(100);
+						$isrunning = false;
+						break;
+					}
+					fflush(STDOUT);
+					usleep(10000);
+				}
+
+				$progress->finish();
+				// Re-block the stream
+				if (!$output->isQuiet()) {
+					stream_set_blocking(STDIN,1);
+					system("stty $term");
+				}
+
+				if ($isrunning) {
+					$output->writeln("");
+					$output->writeln(_('Killing asterisk forcefully.'));
+					$this->stopAsterisk($output, 'now');
 				}
 			}
 		}
+		$output->writeln("");
+
+		if ($runpost) {
+			foreach($post as $pri => $data) {
+				$bmo->$data['module']->$data['method']($output);
+			}
+		}
 	}
+
 	private function asteriskProcess(){
 		$ps = '/usr/bin/env ps';
 		$cmd = $ps . " -C asterisk --no-headers -o '%p|%t'";
@@ -164,17 +199,17 @@ class Stop extends Command {
 		return explode('|',$stat);
 	}
 	private function stopAsterisk($output, $method){
-		$output->writeln(_('Stopping Asterisk...'));
-		$sastbin = '/usr/bin/env killall safe_asterisk > /dev/null 2>&1';
-		$astbin = '/usr/bin/env asterisk -rx "core stop ' . $method .'"';
-		exec($sastbin);
-		exec($astbin);
+		if ($method === "now") {
+			$sastbin = '/usr/bin/env killall safe_asterisk > /dev/null 2>&1';
+			exec($sastbin);
+		}
+		$astbin = '/usr/bin/env asterisk -rx "core stop ' . $method .'" &>/dev/null &';
+		shell_exec($astbin);
 	}
 	private function abortShutdown($output){
 		$freepbx = \FreePBX::Create();
 		$astman = $freepbx->astman;
 		if (is_object($astman) && $astman->Connected()) {
-			$output->writeln(_('Aborting Shutdown'));
 			$astman->send_request('Command',array('Command'=>'core abort shutdown'));
 		}
 	}
@@ -191,5 +226,20 @@ class Stop extends Command {
 		}
 		\FreePBX::Hooks()->processHooks($output);
 		return;
+	}
+
+	private function showHelp() {
+		$help = "<info>"._('Usage').": fwconsole stop [--immediate|--pre|--post] [-m..|--maxwait=..] [modulename] [modulename...]</info>".PHP_EOL;
+		$options = array(
+			"--immediate" => _("Run an immediate shutdown. Defaults to 'when convenient'"),
+			"--pre" => _("Force run pre-stop asterisk hooks"),
+			"--post" => _("Force run post-stop asterisk hooks"),
+			"--maxwait" => _("Maximum amount of time (in seconds) to wait for asterisk to shut down"),
+		);
+		foreach ($options as $o => $t) {
+			$help .= "<info>$o</info> : <comment>$t</comment>".PHP_EOL;
+		}
+
+		return $help;
 	}
 }
